fix: Q_hat now gets proper values for s and s_prime
Reviewer 1 keyed me in on the fact that my start_state and end_state
may end up being incorrect values, because the step function
doesn't necessarily get called until after an update.  By moving
the Q_hat to the start of the update function, and moving the
previous state, reward, and action out of the function, Q_hat
is guaranteed to have the correct states.